Remove dependency of data table repos on data-table-rows.repository
